got to local and remote comparison

+++ lib/code_behind/google_drive/online_sync_state.dart
@@ -9,7 +9,7 @@ class OnlineSyncState {
   final OnlineSyncStateEnum state;
 
   /// in percent
-  final double? progress;
+  final int? progress;
   final String? errorMsg;
 
   OnlineSyncState({
